+++ chainservice/builder.go
@@ -246,7 +246,7 @@ func (builder *Builder) buildBlockDAO(forTest bool) error {
 	}
 
 	var indexers []blockdao.BlockIndexer
-	indexers = append(indexers, builder.cs.factory, builder.cs.liquidStakingIndexer)
+	indexers = append(indexers, builder.cs.factory)
 	if !builder.cfg.Chain.EnableAsyncIndexWrite && builder.cs.indexer != nil {
 		indexers = append(indexers, builder.cs.indexer)
 	}
@@ -266,7 +266,7 @@ func (builder *Builder) buildBlockDAO(forTest bool) error {
 }
 
 func (builder *Builder) buildGatewayComponents(forTest bool) error {
-	indexer, bfIndexer, candidateIndexer, candBucketsIndexer, liquidStakeBucketsIndexer, err := builder.createGateWayComponents(forTest)
+	indexer, bfIndexer, candidateIndexer, candBucketsIndexer, err := builder.createGateWayComponents(forTest)
 	if err != nil {
 		return errors.Wrapf(err, "failed to create gateway components")
 	}
@@ -280,10 +280,6 @@ func (builder *Builder) buildGatewayComponents(forTest bool) error {
 	}
 	builder.cs.bfIndexer = bfIndexer
 	builder.cs.indexer = indexer
-	builder.cs.liquidStakingIndexer = liquidStakeBucketsIndexer
-	// if builder.cs.liquidStakingIndexer != nil {
-	// 	builder.cs.lifecycle.Add(builder.cs.liquidStakingIndexer)
-	// }
 	return nil
 }
 
@@ -292,7 +288,6 @@ func (builder *Builder) createGateWayComponents(forTest bool) (
 	bfIndexer blockindex.BloomFilterIndexer,
 	candidateIndexer *poll.CandidateIndexer,
 	candBucketsIndexer *staking.CandidatesBucketsIndexer,
-	liquidStakeBucketsIndexer blockindex.LiquidStakingIndexer,
 	err error,
 ) {
 	_, gateway := builder.cfg.Plugins[config.GatewayPlugin]
@@ -316,7 +311,6 @@ func (builder *Builder) createGateWayComponents(forTest bool) (
 		if builder.cfg.Chain.EnableStakingIndexer {
 			candBucketsIndexer, err = staking.NewStakingCandidatesBucketsIndexer(db.NewMemKVStore())
 		}
-		liquidStakeBucketsIndexer = blockindex.NewLiquidStakingIndexer(db.NewMemKVStore(), builder.cfg.Genesis.BlockInterval)
 		return
 	}
 	dbConfig := builder.cfg.DB
@@ -345,10 +339,6 @@ func (builder *Builder) createGateWayComponents(forTest bool) (
 		dbConfig.DbPath = builder.cfg.Chain.StakingIndexDBPath
 		candBucketsIndexer, err = staking.NewStakingCandidatesBucketsIndexer(db.NewBoltDB(dbConfig))
 	}
-
-	// create liquid staking indexer
-	dbConfig.DbPath = builder.cfg.Chain.LiquidStakingIndexDBPath
-	liquidStakeBucketsIndexer = blockindex.NewLiquidStakingIndexer(db.NewBoltDB(dbConfig), builder.cfg.Genesis.BlockInterval)
 	return
 }
