+++ chainservice/chainservice.go
@@ -80,13 +80,12 @@ type ChainService struct {
 	p2pAgent          p2p.Agent
 	electionCommittee committee.Committee
 	// TODO: explorer dependency deleted at #1085, need to api related params
-	indexer              blockindex.Indexer
-	bfIndexer            blockindex.BloomFilterIndexer
-	candidateIndexer     *poll.CandidateIndexer
-	candBucketsIndexer   *staking.CandidatesBucketsIndexer
-	registry             *protocol.Registry
-	nodeInfoManager      *nodeinfo.InfoManager
-	liquidStakingIndexer blockindex.LiquidStakingIndexer
+	indexer            blockindex.Indexer
+	bfIndexer          blockindex.BloomFilterIndexer
+	candidateIndexer   *poll.CandidateIndexer
+	candBucketsIndexer *staking.CandidatesBucketsIndexer
+	registry           *protocol.Registry
+	nodeInfoManager    *nodeinfo.InfoManager
 }
 
 // Start starts the server
